requirements.txt
Flask==3.0.2
python-chess==1.999
